requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.3.0
pydantic==2.11.7
email-validator==2.3.0
